Clean up solver interface
For early debugging/development, the solver just printed out its results
at the end. Work towards a better interface by having the solver report
its progress and results programmatically. It now reports:
- whether it was successful in solving taxes
- the solution as a ConfigParser object (partial solution if
  unsuccessful)
- unmet input dependencies if unsuccessful
- unmet field dependencies if unsuccessful
- fields encountering unimplemented behavior if unsuccessful

This also adds handling to __main__.py to print the results and/or unmet
dependencies, or to write them to an output file.