solver: Make input prompts be callbacks
This introduces a cleaner separation of concerns between the solver
kernel and the UI (and would allow for multiple user-interfaces to use
the same solver at some point in the future if desired).

This change necessitated moving the Input objects to accept string input
rather than read their values directly from a ConfigParser. This is a
better design, anyway, because it more cleanly encapsulates the inputs
and doesn't require them to understand details of where their values are
coming from or how to get them.